backend/project/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
Django
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
simplejson==3.19.2
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
django-cors-headers==4.0.0